+++ types/physX.d.ts
@@ -1,0 +1,6 @@
+/**
+ * dom元素是否相交（碰撞）
+ * @param r1
+ * @param r2
+ */
+export declare function haveDomRectIntersection(r1: DOMRect, r2: DOMRect): boolean;